db-monitoring-project/db-loader/requirements.txt
psycopg2-binary
prometheus_client
